% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{demo1}
\alias{demo1}
\title{Run a shiny app to plot funnels or sliced funnels from uploaded mortality data}
\usage{
demo1()
}
\description{
Run a shiny app to plot funnels or sliced funnels from uploaded mortality data
}

