% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnel4.R
\name{funnel4}
\alias{funnel4}
\title{Core funnel and slice plot generator}
\usage{
funnel4(obs.prop = NA, denom = NA, pred.prop = NA, names = NA,
  plot = "funnel", rank = "none", riskadj = F, RASRplot = F,
  ratedenom = NA, mean.target = T, target = NA, plot.target = F,
  title = "", scale = 0.7, xrange = c(0, 1000), yrange = c(0, 1),
  tails = c(0.001, 0.025), Npoints = 200, xlab = "", ylab = "",
  pointsymbol = 16, legend = 1, ypercent = T, bandcols = c("white",
  "lightblue", "azure"))
}
\arguments{
\item{obs.prop}{A data frame of mortality observations from a number of hospitals.}

\item{denom}{Number of cases}
}
\description{
Core funnel and slice plot generator
}

