% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run a shiny app to plot funnels or sliced funnels from uploaded mortality data}
\usage{
run_app()
}
\description{
Run a shiny app to plot funnels or sliced funnels from uploaded mortality data
}

